Move Expression class to expression.cpp/h
