+++ pol-core/bscript/CMakeSources.cmake
@@ -36,6 +36,8 @@ set (bscript_sources    # sorted !
   executor.cpp
   executor.h
   executortype.h
+  expression.cpp
+  expression.h
   filefmt.h
   fmodule.cpp
   fmodule.h
